+++ assemblers/tcp_stream.go
@@ -161,6 +161,9 @@ func (t *tcpStream) ReassemblyComplete(ac reassembly.AssemblerContext) bool {
 		Str("tcp_stream_ident", t.ident).
 		Msg("Connection closed")
 	t.close()
+
+	// decrement the number of active streams
+	DecrementActiveStreamCount()
 	return true // remove the connection, heck with the last ACK
 }